pyproject.toml
[tool.poetry]
name = "hw-7-backend-yersmad"
version = "0.1.0"
description = ""
authors = ["yersmad <yersmad@gmail.com>"]
readme = "README.md"
packages = [{include = "hw_7_backend_yersmad"}]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.23.2"
fastapi = "^0.101.0"
python-multipart = "^0.0.6"
python-jose = "^3.3.0"
jinja2 = "^3.1.2"
attrs = "^23.1.0"
black = "^23.7.0"
requests = "^2.31.0"
requires = "^0.10.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
